requirements.txt
Django == 3.2.7
PyYAML == 5.4.1
django-rest-passwordreset == 1.2.0
djangorestframework == 3.12.4
psycopg2 == 2.9.1
psycopg2-binary == 2.9.1
python-dotenv == 0.19.0
requests == 2.26.0
ujson == 4.1.0
drf-spectacular == 0.18.2
celery == 5.1.2
redis == 3.5.3
pytest == 6.2.5
pytest-django == 4.4.0
django-pagination == 1.0.7